Time: 27 ms (5.24%) | Memory: 22.8 MB (73.93%) - LeetSync

+++ 907-koko-eating-bananas/koko-eating-bananas.cpp
@@ -3,7 +3,7 @@ public:
     long long totalHours(vector<int>& piles, int s){
         long long sum{0};
         if(s<=0) return INT_MAX;
-        for(float i:piles){
+        for(int i:piles){
             sum+=ceil(static_cast<double>(i)/s);
         }
         return sum;
